requirements.txt
boto3
django
djangorestframework
openai
python-dotenv
requests